dist

+++ src/components/Login.vue
@@ -21,6 +21,9 @@ import '../css/Login.css'
 
 import * as API from '../../serve/fontend/index.js'
 import * as BASE from '../../serve/fontend/base.js'
+
+var LOGIN_CODE =  require('../../../login/lib/login.js').CODE
+
 import co from 'co'
 export default {
 		data () {
@@ -54,11 +57,25 @@ export default {
 			
 		}
 	},
-	computed: {
+	computed:{
 
 	},
+	created:function(){
+		let self = this
+		API.LOGIN.login_status_check()
+		.then(function(res){
+			console.log(res)
+			self.$router.push('WriteArticle')
+		})
+		.catch(function(err){
+			console.log(err)
+            if(err.STATUSCODE === LOGIN_CODE.LOGIN_NO_LOGIN.STATUSCODE){
+				localStorage.clear()
+            }
+		})
+	},
 	ready:function(){
-		localStorage.clear()
+
 	}
 }
 </script>
